Bug 1931789. Change DocumentOrShadowRoot::StyleOrderIndexOfSheet to return size_t. r=emilio
size_t is the return type of nsTArray::IndexOf. NoIndex is defined as (size_t)-1. Implicitly converting this to int32_t which is signed and might be a different bit width seems sketchy to me. This makes this clear and explicit.

Differential Revision: https://phabricator.services.mozilla.com/D229282

+++ dom/base/DocumentOrShadowRoot.h
@@ -85,7 +85,7 @@ class DocumentOrShadowRoot {
    *
    * Handles sheets from both mStyleSheets and mAdoptedStyleSheets
    */
-  int32_t StyleOrderIndexOfSheet(const StyleSheet& aSheet) const;
+  size_t StyleOrderIndexOfSheet(const StyleSheet& aSheet) const;
 
   StyleSheetList* StyleSheets();
